requirements.txt
fastapi
uvicorn
python-jose
pydantic
python-dotenv
openai
transformers
torch
pytest
httpx
